+++ src/main/java/ua/beauty/calendar/domain/Procedure.java
@@ -17,10 +17,11 @@ import javax.persistence.Id;
 public class Procedure {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    // private Long id;
-    private Integer id;
+    //  @JsonProperty("id")
+    private Long id;
 
     @NonNull
+    // @JsonProperty("name")
     private String name;
